requirements.txt
django>=4.0,<5.3
djangorestframework
psycopg2-binary
python-dotenv
gunicorn
celery
redis
openpyxl
